[Terry N.] adjust routing and model behavior when the user adds/deletes edges so that results are drawn immediately
git-svn-id: svn+ssh://eos.nps.edu/usr/local/svn/wcm/Viskit/trunk@2539 0de750ab-c0a9-dc11-9d74-0011435b89d6

+++ src/viskit/model/AssemblyModelImpl.java
@@ -45,7 +45,7 @@ public class AssemblyModelImpl extends mvcAbstractModel implements AssemblyModel
     private AssemblyController controller;
 
     public AssemblyModelImpl(AssemblyController cont) {
-        pointLess = new Point2D.Double(10, 10);
+        pointLess = new Point2D.Double(30, 60);
         controller = cont;
         metaData = new GraphMetaData(this);
         setNodeCache(new LinkedHashMap<String, AssemblyNode>());
@@ -94,7 +94,7 @@ public class AssemblyModelImpl extends mvcAbstractModel implements AssemblyModel
     @Override
     public boolean newModel(File f) {
         getNodeCache().clear();
-        pointLess = new Point2D.Double(10, 10);
+        pointLess = new Point2D.Double(30, 60);
         this.notifyChanged(new ModelEvent(this, ModelEvent.NEWASSEMBLYMODEL, "New empty assembly model"));
 
         if (f == null) {
